refactor(LIVE-30363): move @features/market-banner under features/flow/
Renames @features/market-banner to @features/flow-market-banner and moves
it from features/ to features/flow/ per the Iteration 5 Data Layer ADR.
Adds features/flow/* to pnpm-workspace.yaml so pnpm picks up the nested
path.
