+++ .changeset/move-market-banner-to-features-flow.md
@@ -1,0 +1,7 @@
+---
+"@features/flow-market-banner": minor
+"ledger-live-desktop": minor
+"live-mobile": minor
+---
+
+Move `@features/market-banner` to `@features/flow-market-banner` under `features/flow/` per the Iteration 5 Data Layer ADR. [LIVE-30363](https://ledgerhq.atlassian.net/browse/LIVE-30363)
