+++ apps/ledger-live-desktop/package.json
@@ -61,7 +61,7 @@
     "@datadog/browser-rum": "6.30.1",
     "@braze/web-sdk": "6.4.0",
     "@electron/fuses": "2.0.0",
-    "@features/market-banner": "workspace:^",
+    "@features/flow-market-banner": "workspace:^",
     "@ledgerhq/asset-aggregation": "workspace:^",
     "@ledgerhq/asset-detail": "workspace:^",
     "@ledgerhq/client-ids": "workspace:*",
