+++ apps/ledger-live-desktop/tools/rspack/rspack.common.ts
@@ -1,35 +1,9 @@
 import path from "path";
-import fs from "fs";
 import type { RspackOptions } from "@rspack/core";
 
 export const rootFolder = path.resolve(__dirname, "..", "..");
 export const srcFolder = path.resolve(rootFolder, "src");
 export const outputFolder = path.resolve(rootFolder, ".webpack");
-export const featuresFolder = path.resolve(rootFolder, "..", "..", "features");
-
-/**
- * Build dynamic aliases for all features in the features folder.
- * Maps @features/<name> to features/<name>/src for each feature.
- */
-function buildFeaturesAliases(): Record<string, string> {
-  const aliases: Record<string, string> = {};
-
-  if (!fs.existsSync(featuresFolder)) {
-    return aliases;
-  }
-
-  const entries = fs.readdirSync(featuresFolder, { withFileTypes: true });
-  for (const entry of entries) {
-    if (entry.isDirectory()) {
-      const featureSrcPath = path.join(featuresFolder, entry.name, "src");
-      if (fs.existsSync(featureSrcPath)) {
-        aliases[`@features/${entry.name}`] = featureSrcPath;
-      }
-    }
-  }
-
-  return aliases;
-}
 
 /**
  * Common rspack configuration shared across all build targets
@@ -46,8 +20,6 @@ export const commonConfig: RspackOptions = {
     extensions: [".web.tsx", ".web.ts", ".tsx", ".ts", ".jsx", ".js", ".json", ".lottie"],
     alias: {
       "~": srcFolder,
-      // @features/* aliases are dynamically generated for each feature
-      ...buildFeaturesAliases(),
     },
   },
   module: {
