+++ features/flow/market-banner/src/components/MarketBanner/__tests__/MarketBanner.native.test.tsx
