+++ features/flow/market-banner/src/components/MarketBanner/index.native.ts
