+++ features/flow/market-banner/src/utils/index.ts
